+++ src/agentlang/cli/util.clj
@@ -6,6 +6,10 @@
 (set! *warn-on-reflection* true)
 
 
+(defn throw-ex-info [message data]
+  (throw (ex-info message data)))
+
+
 (defmacro retry-if-interrupted
   [& body]
   `(loop []
@@ -104,3 +108,17 @@
     (cond
       (.isDirectory f) "directory"
       (.isFile f) "file")))
+
+
+(def file-separator File/separator)
+(def path-separator File/pathSeparator)
+
+
+(defn git-repo-uri->repo-name [repo-uri]
+  (let [last-name (-> repo-uri
+                      (string/split #"/")
+                      last)]
+    (if (string/ends-with? last-name ".git")
+      (subs last-name 0
+            (- (count last-name) 4))
+      last-name)))